backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.30.0
bcrypt==5.0.0
click==8.3.3
fastapi==0.115.6
h11==0.16.0
httptools==0.7.1
idna==3.13
Jinja2==3.1.4
MarkupSafe==3.0.3
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.12.1
python-dotenv==1.0.1
PyYAML==6.0.3
starlette==0.41.3
typing_extensions==4.15.0
ua-parser==1.0.2
ua-parser-builtins==202603
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
